feat(6948): extract block pipeline modules
