+++ crates/kamn-core/src/block_pipeline/commit_hooks.rs
@@ -1,0 +1,9 @@
+mod pipeline;
+mod replay;
+mod sorting;
+mod transport_pipeline;
+
+pub use pipeline::MempoolBlockPipeline;
+pub use replay::build_canonical_replay_evidence_bundle;
+pub(crate) use sorting::payload_digest_for_transactions;
+pub use transport_pipeline::TransportFedBlockPipeline;
